@@ -73,10 +73,18 @@ class AssetEntityImageProvider extends ImageProvider<AssetEntityImageProvider> {
 
     if (isOriginal ?? false) {
       print('Loading original...');
-      data = picStore.isPrivate ? await key.picStore.assetOriginBytes : await key.picStore.entity.originBytes;
+      data = picStore.isPrivate
+          ? await key.picStore.assetOriginBytes
+          : await key.picStore.entity.originBytes;
     } else {
       print('Loading thumbnail...');
-      data = picStore.isPrivate ? await key.picStore.assetThumbBytes : await key.picStore.entity.thumbDataWithSize(thumbSize[0], thumbSize[1]);
+      if (picStore.entity == null) {
+        print('Entity is null & isPrivate: ${picStore.isPrivate}');
+      }
+      data = picStore.isPrivate
+          ? await key.picStore.assetThumbBytes
+          : await key.picStore.entity
+              .thumbDataWithSize(thumbSize[0], thumbSize[1]);
     }
 
     // if (picStore.isPrivate == true) {
@@ -105,7 +113,9 @@ class AssetEntityImageProvider extends ImageProvider<AssetEntityImageProvider> {
   /// 并非所有的系统版本都支持读取文件名，所以该方法有时无法返回正确的type。
   ImageFileType _getType() {
     ImageFileType type;
-    final String extension = picStore.entity == null ? picStore.photoPath?.split('.')?.last : picStore.entity.title?.split('.')?.last;
+    final String extension = picStore.entity == null
+        ? picStore.photoPath?.split('.')?.last
+        : picStore.entity.title?.split('.')?.last;
     print('Extension: $extension');
     if (extension != null) {
       switch (extension.toLowerCase()) {
