@@ -1,4 +1,5 @@
 import 'dart:async';
+import 'dart:isolate';
 import 'dart:typed_data';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/foundation.dart';
@@ -37,7 +38,8 @@ import 'dart:io';
 
 Future<String> checkForAppStoreInitiatedProducts() async {
   print('Checking if appstore initiated products');
-  List<IAPItem> appStoreProducts = await FlutterInappPurchase.instance.getAppStoreInitiatedProducts();
+  List<IAPItem> appStoreProducts =
+      await FlutterInappPurchase.instance.getAppStoreInitiatedProducts();
   if (appStoreProducts.length > 0) {
     return appStoreProducts.last.productId;
   }
@@ -48,13 +50,19 @@ void main() async {
   WidgetsFlutterBinding.ensureInitialized();
   GestureBinding.instance.resamplingEnabled = true;
 
-  await Firebase.initializeApp();
-  await FirebaseCrashlytics.instance.setCrashlyticsCollectionEnabled(kDebugMode ? false : true);
-  FlutterError.onError = FirebaseCrashlytics.instance.recordFlutterError;
   // CloudFunctions.instance.useFunctionsEmulator(origin: Platform.isAndroid ? 'http://10.0.2.2:5001' : 'http://localhost:5001');
 
-  Ads.initialize();
-  Ads.loadRewarded();
+  await Firebase.initializeApp();
+  await FirebaseCrashlytics.instance.setCrashlyticsCollectionEnabled(true);
+  // await FirebaseCrashlytics.instance.setCrashlyticsCollectionEnabled(kDebugMode ? false : true);
+  FlutterError.onError = FirebaseCrashlytics.instance.recordFlutterError;
+  Isolate.current.addErrorListener(RawReceivePort((pair) async {
+    final List<dynamic> errorAndStacktrace = pair;
+    await FirebaseCrashlytics.instance.recordError(
+      errorAndStacktrace.first,
+      errorAndStacktrace.last,
+    );
+  }).sendPort);
 
   await Hive.initFlutter();
   Hive.registerAdapter(UserAdapter());
@@ -66,18 +74,55 @@ void main() async {
   var picsBox = await Hive.openBox('pics');
   var tagsBox = await Hive.openBox('tags');
 
-  var secretKey = Uint8List.fromList(
-      [76, 224, 117, 70, 57, 101, 39, 29, 48, 239, 215, 240, 41, 149, 198, 69, 64, 5, 207, 227, 190, 126, 8, 133, 136, 234, 130, 91, 254, 104, 196, 158]);
+  var secretKey = Uint8List.fromList([
+    76,
+    224,
+    117,
+    70,
+    57,
+    101,
+    39,
+    29,
+    48,
+    239,
+    215,
+    240,
+    41,
+    149,
+    198,
+    69,
+    64,
+    5,
+    207,
+    227,
+    190,
+    126,
+    8,
+    133,
+    136,
+    234,
+    130,
+    91,
+    254,
+    104,
+    196,
+    158
+  ]);
   var secretBox = await Hive.openBox('secrets', encryptionKey: secretKey);
 
-  String deviceLocale = await DeviceLocale.getCurrentLocale().then((Locale locale) => locale.toString());
-  String appVersion = await PackageInfo.fromPlatform().then((PackageInfo packageInfo) => packageInfo.version);
+  String deviceLocale = await DeviceLocale.getCurrentLocale()
+      .then((Locale locale) => locale.toString());
+  String appVersion = await PackageInfo.fromPlatform()
+      .then((PackageInfo packageInfo) => packageInfo.version);
   print('Device Locale: $deviceLocale');
 
   String initiatedWithProduct;
   if (Platform.isIOS) {
     initiatedWithProduct = await checkForAppStoreInitiatedProducts();
   }
+
+  Ads.initialize();
+  Ads.loadRewarded();
 
   runZonedGuarded<Future<void>>(() async {
     runApp(
